Refactor code and update dependencies
This commit applies some refactoring to the codebase and updates dependencies in the Cargo.toml file. Code readability is improved by re-organizing the structure of syntax within modules and refactoring some of the function declarations to make them more concise. Additionally, the versions of the `wgpu` and `winit` dependencies have been specified.

Signed-off-by: xxorza <1442523+xorza@users.noreply.github.com>