Update dependencies and enhance mouse functionality
This update increases the versions of wgpu and winit in the dependencies, thereby facilitating an upgrade in functionality. The event handling for mouse buttons has been enriched by including Back and Forward buttons. The runtime has been refactored to improve error handling and surface configuration logic. Various changes were also made to comply with the updated wgpu and winit libraries.

Signed-off-by: xxorza <1442523+xorza@users.noreply.github.com>