Refactor wgpu_app.rs to separate surface and window from AppContext
The Window and Surface properties have been removed from the AppContext in order to make the state management clearer and more manageable. They are now directly under AppState which improves the structural design of the application. The Window and Surface are now being passed through the application where necessary by referencing the new AppState properties.

Signed-off-by: xxorza <1442523+xorza@users.noreply.github.com>