Refactor AppState struct to use Arc for window instance
Institute use of Arc for main window in the AppState struct. Simplify the AppState and AppContext structs by moving window and surface ownership to AppContext. Remove instances of calls to 'unwrap()' on main window and surface, which increase safety by eliminating potential panic points. This refactoring improves the code organization and safety.

Signed-off-by: xxorza <1442523+xorza@users.noreply.github.com>