Update dependencies and configurations
The dependencies in the 'Cargo.toml' have been updated to the latest versions. This includes switching the source of 'imaginarium' from a Git URL to a local workspace. Some non-essential build flags have been removed from the run command for the 'the_matrix' package, and an additional directory has been added to the 'excludeFolder' list in the 'wgpu_app.iml' file.